## Description

> Known issue:

> 1. In chinese inference case client backend, chatglm2 with a long max-length(>128) works really slow in my machine.
> 2. Chatglm2 quantization takes about 8G memory and cpu resource, which present a terrible stock in my client machine.
> 3. Tranditional Chinese presented in langchain, and unstable output in Chatglm2:

1. has been fixed
2. fixed, Could be fixed by load int4 model with `load_low_bit` API in TransformersLLM.from_model_id. Note that you need to save transformer int4 model in advance:
```python
from bigdl.llm.transformers import AutoModel

model = AutoModel.from_pretrained("./chatglm2",
                                             load_in_4bit=True),
model.save_pretrained("./chatglm2-int4")
```
4. pending, but englist output looks good and stable

load int4 model with out quant:
```python
 python .\langchain_whisper.py -m .\chatglm2-int4\ -r .\whisper-tiny\ -l english -d
```
output:
```log
> Entering new  chain...
a call is called
Prompt after formatting:


Q:  What is there?
A:
C:\Users\zhaoc1\AppData\Local\anaconda3\envs\bigdl-llm\lib\site-packages\transformers\generation\utils.py:1259: UserWarning: You have modified the pretrained model configuration to control generation. This is a deprecated strategy to control generation and will be removed soon, in a future version. Please use a generation configuration file (see https://huggingface.co/docs/transformers/main_classes/text_generation)
  warnings.warn(
The attention mask and the pad token id were not set. As a consequence, you may observe unexpected behavior. Please pass your input's `attention_mask` to obtain reliable results.
Setting `pad_token_id` to `eos_token_id`:2 for open-end generation.

> Finished chain.
 There is nothing. There is no matter what you want there to be.
Recognized in 0.5688877105712891s, Predicted in 2.4757232666015625s
listening now...
Recognizing...

> Entering new  chain...
a call is called
Prompt after formatting:

Human:  What is there?
AI:  There is nothing. There is no matter what you want there to be.
Q:  introduce yourself
A:

> Finished chain.
  Hello, my name is Q. I am an artificial intelligence and I am here to assist you with any questions or tasks you may have.
Recognized in 0.4364185333251953s, Predicted in 5.290874481201172s
```